making edits so nested JSON can be read
