+++ MemberTracker/Program.cs
@@ -1,6 +1,5 @@
 ﻿using MemberTracker.Clients;
 using Common.Utils;
-using Common.Config;
 
 namespace MemberTracker
 {
@@ -8,8 +7,12 @@ namespace MemberTracker
     {
         public static async Task Main(string[] args)
         {
+            // get secrets
             var secrets = new Config(Constants.SECRETS_FILEPATH);
-            var proPublicaClient = new ProPublicaClient(secrets.Properties["proPublicaApiKey"]);
+            string apiKey = secrets.Properties.proPublicaApiKey;
+
+            // call API and get congressional data
+            var proPublicaClient = new ProPublicaClient(apiKey);
             var houseMembers = await proPublicaClient.GetHouseMembersAsync();
             var senateMembers = await proPublicaClient.GetSenateMembersAsync();
         }
